No argmin implementation for bool - should add a test for this

+++ fastmri/models/varnet.py
@@ -182,7 +182,7 @@ class SensitivityModel(nn.Module):
     ) -> Tuple[torch.Tensor, torch.Tensor]:
         if num_low_frequencies is None:
             # get low frequency line locations and mask them out
-            squeezed_mask = mask[:, 0, 0, :, 0]
+            squeezed_mask = mask[:, 0, 0, :, 0].to(torch.int8)
             cent = squeezed_mask.shape[1] // 2
             # running argmin returns the first non-zero
             left = torch.argmin(squeezed_mask[:, :cent].flip(1), dim=1)
